refactor: vo 패키지 생성, 'Lotto', 'Money' vo 패키지 이동
