+++ src/main/java/lotto/model/LottoStore.java
@@ -3,6 +3,8 @@ package lotto.model;
 import java.util.ArrayList;
 import java.util.Comparator;
 import java.util.List;
+import lotto.model.vo.Lotto;
+import lotto.model.vo.Money;
 
 public class LottoStore {
